requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
google-generativeai==0.5.4
reportlab==4.1.0
pypdf2==3.0.1
PyMuPDF==1.24.2
python-dateutil==2.9.0
